requirements.txt
streamlit==1.16.0
scipy==1.10.0
st-click-detector==0.1.3
st-clickable-images==0.0.3
numpy==1.24.1 	
opencv-python==4.7.0.68
Pillow==9.4.0
gdown==4.6.0
opencv-contrib-python==4.7.0.68